dalle-2/pyproject.toml
[tool.poetry]
name = "dalle-2"
version = "0.1.0"
description = ""
authors = ["KotaroYamazaki <sentoryoku53m@gmail.com>"]
readme = "README.md"
packages = [{include = "dalle_2"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.12.0"
requests = "^2.31.0"
DateTime = "^5.4"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
